requirements.txt
# TF-DocGen dependencies
# Pinned to exact versions for reproducibility and supply-chain security.
# To upgrade: test against the new version, then update the pin here.
python-hcl2==4.3.4
jinja2==3.1.5
click==8.1.8
